Initial support for line segment start/end caps
Relates to #55

D2D supports various line caps. This commit exposes those to users of the API so that values other than "flat" can be specified.

There are other draw methods for which this feature could be enabled (such as poly lines) that are not covered here.

This commit adds optional parameters to public API which is a breaking binary change, but a non-breaking source change.